+++ opentelemetry-sdk/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- Make `SpanProcessor.on_start` accept parent Context
+  ([#1251](https://github.com/open-telemetry/opentelemetry-python/pull/1251))
+- Fix b3 propagator entrypoint
+  ([#1265](https://github.com/open-telemetry/opentelemetry-python/pull/1265))
+
 ## Version 0.14b0
 
 Released 2020-10-13